feat: create Maincontent
